+++ Sources/EmojiKit/EmojiCategory+Persisted.swift
@@ -22,14 +22,20 @@ public extension EmojiCategory {
         ///
         /// - Parameters:
         ///   - id: The category ID.
+        ///   - insertionStrategy: The insertion strategy to use.
         public init(
-            id: String
+            id: String,
+            insertionStrategy: EmojiInsertionStrategy
         ) {
             self.id = id
+            self.insertionStrategy = insertionStrategy
         }
         
         /// The category ID.
         public let id: String
+        
+        /// The insertion strategy to use.
+        public let insertionStrategy: EmojiInsertionStrategy
         
         /// The store to use.
         public var store: UserDefaults { .standard }
@@ -46,38 +52,43 @@ public extension EmojiCategory {
 
 public extension EmojiCategory.Persisted {
     
-    /// A persisted category for favorite emojis.
-    static let favorites = Self.init(id: "favorites")
-    
-    /// A persisted category for frequently used emojis.
-    static let frequent = Self.init(id: "frequent")
-    
-    /// A persisted category for recent emojis.
-    static let recent = Self.init(id: "recent")
+    /// A strategy that defines how new emojis are added.
+    enum EmojiInsertionStrategy: Codable, Equatable, Sendable {
+        
+        /// Append new emojis last in the category.
+        case append
+        
+        /// Insert new emojis first in the category.
+        case insertFirst
+    }
 }
 
 public extension EmojiCategory.Persisted {
     
-    /// The max number of emojis to store in the category.
-    var emojisMaxCount: Int {
-        get {
-            let key = storageKey(for: .emojisMaxCount)
-            let value = store.integer(forKey: key)
-            return value < 1 ? 1_000 : value
-        }
-        set {
-            let key = storageKey(for: .emojisMaxCount)
-            return store.set(newValue, forKey: key)
-        }
-    }
+    /// A persisted category for favorite emojis.
+    static let favorites = Self.init(
+        id: "favorites",
+        insertionStrategy: .append
+    )
+    
+    /// A persisted category for recent emojis.
+    static let recent = Self.init(
+        id: "recent",
+        insertionStrategy: .insertFirst
+    )
+}
+
+public extension EmojiCategory.Persisted {
     
     /// Add an emoji to the category.
     func addEmoji(
         _ emoji: Emoji
     ) {
-        var emojis = getEmojis()
-            .filter { $0 != emoji }
-        emojis.insert(emoji, at: 0)
+        var emojis = getEmojis().filter { $0 != emoji }
+        switch insertionStrategy {
+        case .append: emojis.append(emoji)
+        case .insertFirst: emojis.insert(emoji, at: 0)
+        }
         setEmojis(emojis)
     }
     
@@ -88,6 +99,13 @@ public extension EmojiCategory.Persisted {
         return value.map { Emoji($0) }
     }
     
+    /// Get the max number of emojis to persist.
+    func getEmojisMaxCount() -> Int {
+        let key = storageKey(for: .emojisMaxCount)
+        let value = store.integer(forKey: key)
+        return value < 1 ? 1_000 : value
+    }
+    
     /// Remove an emoji from the category.
     func removeEmoji(
         _ emoji: Emoji
@@ -96,9 +114,20 @@ public extension EmojiCategory.Persisted {
         setEmojis(emojis)
     }
     
+    /// Reset the category.
+    func reset() {
+        resetEmojis()
+        resetEmojisMaxCount()
+    }
+    
     /// Reset the emojis in the category.
     func resetEmojis() {
         setEmojis([])
+    }
+    
+    /// Reset the emojis max count for the category.
+    func resetEmojisMaxCount() {
+        setEmojisMaxCount(0)
     }
 
     /// Set the persisted emojis for the category.
@@ -107,13 +136,33 @@ public extension EmojiCategory.Persisted {
         applyMaxCount limit: Bool = true
     ) {
         let key = storageKey(for: .emojis)
-        let emojiChars = emojis.map { $0.char }
-        let value = limit ? Array(emojiChars.prefix(emojisMaxCount)) : Array(emojiChars)
-        return store.set(value, forKey: key)
+        let emojis = limitEmojisForStorage(emojis)
+        let chars = emojis.map { $0.char }
+        return store.set(chars, forKey: key)
+    }
+    
+    /// Set the max number of emojis to persist.
+    func setEmojisMaxCount(_ val: Int) {
+        let key = storageKey(for: .emojisMaxCount)
+        return store.set(val, forKey: key)
     }
 }
 
 private extension EmojiCategory.Persisted {
+    
+    /// Limit a collection before persisting.
+    func limitEmojisForStorage(
+        _ emojis: [Emoji],
+        applyMaxCount limit: Bool = true
+    ) -> [Emoji] {
+        guard limit else { return emojis }
+        let maxCount = getEmojisMaxCount()
+        guard maxCount > 0 else { return emojis }
+        switch insertionStrategy {
+        case .append: return Array(emojis.suffix(maxCount))
+        case .insertFirst: return Array(emojis.prefix(maxCount))
+        }
+    }
     
     /// Get a storage key for a certain value.
     func storageKey(